{{cookiecutter.project_slug}}/requirements/development.txt
-r testing.txt

django-debug-toolbar==3.2
